refactor(apps/web): update player store imports for consistency
- Refactored import paths for `usePlayerStore` across multiple components and tests to use the new `player-store` directory structure.
- Ensured all references to the player store are consistent, improving code maintainability and clarity.
