+++ apps/web/src/components/app/player/PlayerActions.tsx
@@ -14,7 +14,7 @@ import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover
 import { Slider } from '@/components/ui/slider';
 import { listPlaybackDevices, setActivePlaybackDevice } from '@/lib/playback-sync';
 import { cn } from '@/lib/utils';
-import { usePlayerStore } from '@/stores/player.store';
+import { usePlayerStore } from '@/stores/player-store/player.store';
 import {
   CheckIcon,
   DotsThreeIcon,
