+++ apps/web/src/components/app/player/use-player-audio.ts
@@ -1,7 +1,7 @@
 import { getOrderedNextQueue } from '@/lib/playback-queue';
 import { emitCurrentTimeSync, isPlaybackSyncConnected } from '@/lib/playback-sync';
 import { useAuthStore } from '@/stores/auth.store';
-import { usePlayerStore } from '@/stores/player.store';
+import { usePlayerStore } from '@/stores/player-store/player.store';
 import { StreamAudioQuality } from '@repo/contracts';
 import { useEffect, useRef } from 'react';
