+++ apps/web/src/routes/app/library/albums/$id/index.tsx
@@ -21,7 +21,7 @@ import { useDeleteLibraryAlbum } from '@/hooks/api/library-albums/useDeleteLibra
 import { useLibraryAlbum } from '@/hooks/api/library-albums/useLibraryAlbum';
 import { useDeleteLibraryTrack } from '@/hooks/api/library-tracks/useDeleteLibraryTrack';
 import { zodTrackToPlaybackTrack } from '@/lib/playback-mappers';
-import { usePlayerStore } from '@/stores/player.store';
+import { usePlayerStore } from '@/stores/player-store/player.store';
 import {
   DiscIcon,
   DotsThreeIcon,
